Add 'input' event for HTMLInputElement
